pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[project]
name = "broker-buddy-mcp"
version = "0.1.0"
description = "FastMCP server exposing Angel One Smart API tools"
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "Win" }]
license = { text = "MIT" }
keywords = ["mcp", "fastmcp", "angelone", "smartapi"]
dependencies = [
  "attrs",
  "autobahn",
  "automat",
  "certifi",
  "cffi",
  "charset-normalizer",
  "constantly",
  "cryptography",
  "hyperlink",
  "idna",
  "incremental",
  "isodate",
  "logzero",
  "pycparser",
  "pyotp",
  "pyparsing",
  "python-dateutil",
  "pytz",
  "rdflib",
  "rdflib-jsonld",
  "requests",
  "simplejson",
  "six",
  "smartapi-python",
  "twisted",
  "txaio",
  "typing-extensions",
  "urllib3",
  "websocket-client",
  "zope-interface",
  "fastmcp",
  "python-dotenv",
  "fastapi>=0.116.1",
  "coinbase-advanced-py>=1.8.2",
]

[project.urls]
Homepage = "https://example.com"

[project.scripts]
broker-buddy-mcp = "mcp_src.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_src"]

[dependency-groups]
dev = [
    "autoflake>=2.3.1",
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=6.0.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
]


[tool.isort]
profile = "black"
skip = ["SmartApi"]
